+++ src/fhapi/mouse.c
@@ -1,7 +1,7 @@
 /*
 #   clove
 #
-#   Copyright (C) 2019 Muresan Vlad
+#   Copyright (C) 2019-2026 Muresan Vlad
 #
 #   This project is free software; you can redistribute it and/or modify it
 #   under the terms of the MIT license. See LICENSE.md for details.
@@ -12,7 +12,7 @@
 #include "../include/mouse.h"
 
 #include "../3rdparty/FH/src/value.h"
-#include "../3rdparty/SDL2/include/SDL.h"
+#include "../3rdparty/SDL3/include/SDL3/SDL.h"
 
 static struct {
     struct fh_program *prog;