+++ src/include/joystick.h
@@ -2,7 +2,7 @@
 /*
 #   clove
 #
-#   Copyright (C) 2016-2020 Muresan Vlad
+#   Copyright (C) 2016-2026 Muresan Vlad
 #
 #   This project is free software; you can redistribute it and/or modify it
 #   under the terms of the MIT license. See LICENSE.md for details.
@@ -10,13 +10,13 @@
 
 #pragma once
 
-#include "../3rdparty/SDL2/include/SDL.h"
+#include "../3rdparty/SDL3/include/SDL3/SDL.h"
 #include "stdbool.h"
 
 typedef struct {
     SDL_Joystick *joystick;
     SDL_JoystickID id;
-    SDL_GameController *controller;
+    SDL_Gamepad *controller;
 } joystick_Joystick;
 
 // events