+++ src/include/keyboard.h
@@ -1,14 +1,14 @@
 /*
 #   clove
 #
-#   Copyright (C) 2016-2020 Muresan Vlad
+#   Copyright (C) 2016-2026 Muresan Vlad
 #
 #   This project is free software; you can redistribute it and/or modify it
 #   under the terms of the MIT license. See LICENSE.md for details.
 */
 #pragma once
 
-#include "../3rdparty/SDL2/include/SDL_keyboard.h"
+#include "../3rdparty/SDL3/include/SDL3/SDL_keyboard.h"
 
 #include <stdbool.h>