+++ src/luaapi/keyboard.h
@@ -1,7 +1,7 @@
 /*
 #   clove
 #
-#   Copyright (C) 2016-2017 Muresan Vlad
+#   Copyright (C) 2016-2026 Muresan Vlad
 #
 #   This project is free software; you can redistribute it and/or modify it
 #   under the terms of the MIT license. See LICENSE.md for details.
@@ -10,7 +10,7 @@
 
 #ifdef USE_LUA
 #include "../3rdparty/lua/lua.h"
-#include "../3rdparty/SDL2/include/SDL_keycode.h"
+#include "../3rdparty/SDL3/include/SDL3/SDL_keycode.h"
 
 #include <stdbool.h>